ogre_helpers: introduce abstract QtOgreRenderWindow class
