+++ src/rviz/CMakeLists.txt
@@ -63,6 +63,7 @@ add_library(${PROJECT_NAME}
   ogre_helpers/orthographic.cpp
   ogre_helpers/point_cloud.cpp
   ogre_helpers/qt_ogre_render_window.cpp
+  ogre_helpers/qt_widget_ogre_render_window.cpp
   ogre_helpers/render_system.cpp
   ogre_helpers/render_widget.cpp
   ogre_helpers/shape.cpp
