+++ src/rviz/ogre_helpers/qt_ogre_render_window.h
@@ -47,37 +47,26 @@ class Camera;
 namespace rviz
 {
 
-/** Qt Ogre render window widget.  Similar in API to
+/**
+ * \brief Generic interface for Qt Ogre render windows
+ * Qt Ogre render window widget.  Similar in API to
  *  wxOgreRenderWindow from ogre_tools release 1.6, but with much of
  *  the guts replaced by new RenderSystem and RenderWidget classes
  *  inspired by the initialization sequence of Gazebo's renderer.
  */
-class QtOgreRenderWindow : public RenderWidget, public Ogre::RenderTargetListener {
+class QtOgreRenderWindow : public Ogre::RenderTargetListener {
 public:
-  /** Constructor.
-  	@param parent The parent wxWindow component.
-   */
-  QtOgreRenderWindow( QWidget* parent = 0 );
-
-  /** Destructor.  */
-  virtual ~QtOgreRenderWindow();
 
   /**
    * Set a callback which is called before each render
    * @param func The callback functor
    */
-  virtual void setPreRenderCallback( boost::function<void ()> func );
+  virtual void setPreRenderCallback( boost::function<void ()> func ) = 0;
   /**
      * Set a callback which is called after each render
      * @param func The callback functor
      */
-  virtual void setPostRenderCallback( boost::function<void ()> func );
-
-  /** Overrides the default implementation.
-  	This override is here for convenience. Returns a symbolic 320x240px size.
-  	@return A size of 320x240 (just a symbolic 4:3 size).
-   */
-  virtual QSize sizeHint () const { return QSize( 320, 240 ); }
+  virtual void setPostRenderCallback( boost::function<void ()> func ) = 0;
 
   /** Gets the associated Ogre viewport.  If this is called before
    * QWidget::show() on this widget, it will fail an assertion.
@@ -85,77 +74,37 @@ public:
    * which can be called before QWidget::show(), and their effects are
    * propagated to the viewport when it is created.
    */
-  Ogre::Viewport* getViewport() const;
+  virtual Ogre::Viewport* getViewport() const = 0;
+
+  virtual Ogre::RenderWindow* getRenderWindow() = 0;
 
   /** Set the camera associated with this render window's viewport.
    */
-  void setCamera( Ogre::Camera* camera );
+  virtual void setCamera( Ogre::Camera* camera ) = 0;
 
-  Ogre::Camera* getCamera() const { return camera_; }
+  virtual Ogre::Camera* getCamera() const = 0;
 
   /**
    * \brief Set the scale of the orthographic window.  Only valid for an orthographic camera.
    * @param scale The scale
    */
-  void setOrthoScale( float scale );
+  virtual void setOrthoScale( float scale ) = 0;
 
   /** \brief Enable or disable stereo rendering
    * If stereo is not supported this is ignored.
    * @return the old setting (whether stereo was enabled before)
    */
-  bool enableStereo(bool enable);
+  virtual bool enableStereo(bool enable) = 0;
 
   /** \brief Prepare to render in stereo if enabled and supported. */
-  void setupStereo();
+  virtual void setupStereo() = 0;
 
-  void setAutoRender(bool auto_render) { auto_render_ = auto_render; }
+  virtual void setAutoRender(bool auto_render) = 0;
 
   ////// Functions mimicked from Ogre::Viewport to satisfy timing of
   ////// after-constructor creation of Ogre::RenderWindow.
-  void setOverlaysEnabled( bool overlays_enabled );
-  void setBackgroundColor( Ogre::ColourValue color );
-
-protected:
-  virtual void paintEvent( QPaintEvent* e );
-  virtual void resizeEvent( QResizeEvent* event );
-
-  // When stereo is enabled, these are called before/after rendering each
-  // viewport.
-  virtual void preViewportUpdate(const Ogre::RenderTargetViewportEvent& evt);
-  virtual void postViewportUpdate(const Ogre::RenderTargetViewportEvent& evt);
-
-  /**
-   * Sets the aspect ratio on the camera
-   */
-  void setCameraAspectRatio();
-
-  /**
-   * prepare a viewport's camera for stereo rendering.
-   * This should only be called from StereoRenderTargetListener
-   */
-  void prepareStereoViewport(Ogre::Viewport*);
-
-
-  Ogre::Viewport* viewport_;
-
-  Ogre::Root* ogre_root_;
-
-  boost::function<void ()> pre_render_callback_;      ///< Functor which is called before each render
-  boost::function<void ()> post_render_callback_;     ///< Functor which is called after each render
-
-  float ortho_scale_;
-  bool auto_render_;
-
-  Ogre::Camera* camera_;
-  bool overlays_enabled_;
-  Ogre::ColourValue background_color_;
-
-  // stereo rendering
-  bool stereo_enabled_;				// true if we were asked to render stereo
-  bool rendering_stereo_;			// true if we are actually rendering stereo
-  Ogre::Camera* left_camera_;
-  Ogre::Camera* right_camera_;
-  Ogre::Viewport* right_viewport_;
+  virtual void setOverlaysEnabled( bool overlays_enabled ) = 0;
+  virtual void setBackgroundColor( Ogre::ColourValue color ) = 0;
 };
 
 } // namespace rviz
