+++ src/rviz/render_panel.h
@@ -64,6 +64,7 @@ namespace rviz
 class Display;
 class DisplayContext;
 class ViewController;
+class QtOgreRenderWindow;
 
 /**
  * A widget which shows an OGRE-rendered scene in RViz.
@@ -72,7 +73,7 @@ class ViewController;
  * the DisplayContext (which further forwards them to the active
  * Tool, etc.)
  */
-class RenderPanel: public QtOgreRenderWindow, public Ogre::SceneManager::Listener
+class RenderPanel: public QWidget, public Ogre::SceneManager::Listener
 {
 Q_OBJECT
 public:
@@ -86,6 +87,14 @@ public:
   DisplayContext* getManager() { return context_; }
 
   ViewController* getViewController() { return view_controller_; }
+
+  /// Pass through render window functions
+  Ogre::Viewport* getViewport() const { return render_window_->getViewport(); }
+  Ogre::RenderWindow* getRenderWindow() const { return render_window_->getRenderWindow(); }
+  Ogre::Camera* getCamera() const { return render_window_->getCamera(); }
+  void setAutoRender(bool auto_render) { render_window_->setAutoRender(auto_render); }
+  void setBackgroundColor(Ogre::ColourValue color) { render_window_->setBackgroundColor(color); }
+  void setOverlaysEnabled(bool enabled) { render_window_->setOverlaysEnabled(enabled); }
 
   /** @brief Set the ViewController which should control the camera
    * position for this view. */
@@ -146,6 +155,7 @@ private Q_SLOTS:
 private:
   QTimer* fake_mouse_move_event_timer_;
   Ogre::Camera* default_camera_; ///< A default camera created in initialize().
+  QtOgreRenderWindow* render_window_;
 };
 
 } // namespace rviz
