+++ src/rviz/visualization_frame.cpp
@@ -108,18 +108,18 @@ namespace rviz
 
 VisualizationFrame::VisualizationFrame( QWidget* parent )
   : QMainWindow( parent )
-  , app_(NULL)
-  , render_panel_(NULL)
-  , show_help_action_(NULL)
-  , file_menu_(NULL)
-  , recent_configs_menu_(NULL)
-  , toolbar_(NULL)
-  , manager_(NULL)
-  , splash_( NULL )
-  , toolbar_actions_( NULL )
+  , app_( nullptr )
+  , render_panel_( nullptr )
+  , show_help_action_( nullptr )
+  , file_menu_( nullptr )
+  , recent_configs_menu_( nullptr )
+  , toolbar_( nullptr )
+  , manager_( nullptr )
+  , splash_( nullptr )
+  , toolbar_actions_( nullptr )
   , show_choose_new_master_option_( false )
-  , add_tool_action_( NULL )
-  , remove_tool_menu_( NULL )
+  , add_tool_action_( nullptr )
+  , remove_tool_menu_( nullptr )
   , initialized_( false )
   , geom_change_detector_( new WidgetGeometryChangeDetector( this ))
   , loading_( false )
